+++ components/Breadcrumbs.php
@@ -11,7 +11,6 @@
 namespace skeeks\cms\components;
 
 use skeeks\cms\base\components\Descriptor;
-use skeeks\cms\base\db\ActiveRecord;
 use skeeks\cms\models\Site;
 use skeeks\cms\models\StorageFile;
 use skeeks\cms\models\Tree;