+++ modules/admin/traits/AdminModelEditorStandartControllerTrait.php
@@ -8,11 +8,9 @@
 
 namespace skeeks\cms\modules\admin\traits;
 
-use skeeks\cms\base\db\ActiveRecord;
 use skeeks\cms\components\Cms;
 use skeeks\cms\helpers\UrlHelper;
 use skeeks\cms\modules\admin\controllers\AdminModelEditorController;
-use skeeks\cms\validators\db\IsNewRecord;
 use skeeks\sx\validate\Validate;
 use skeeks\widget\chosen\Chosen;
 use yii\base\Model;